pyproject.toml
[tool.poetry]
name = "omoospace"
version = "0.1.1"
description = ""
authors = ["manan <icrdr2010@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ruamel-yaml = "^0.17.35"
pypinyin = "^0.49.0"
click = "^8.1.7"
pyvis = "^0.3.2"
networkx = "^3.1"
inquirerpy = "^0.3.4"
rich = "^13.6.0"
nutree = "^0.5.1"
validators = "^0.22.0"


[tool.poetry.scripts]
omoospace = "omoospace.commands:cli"
omoos = "omoospace.commands:cli"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
pytest = "^7.4.2"
pytest-dependency = "^0.5.1"
mkdocs = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
